Add support for lazy lua value
A lazy value is an expression whose evaluation is delayed. In our context,
it essentially means a value which can be changed after the configuration
file is parsed.

An new lua function is added, `parameter` which accepts a string argument.
Instead of referencing directly the variable in the lua configuration, the
user must use this new function, and pass as argument the name of the
parameter, as specified in the `configuration` table (see updated lua
configuration for an example)