Add a way to edit the configuration before freezing
A new method `set()` is added to `ConfigurationSet` which allow the user
to change a parameter *before* freezing the configuration.

In order to correctly dispatch the change of a parameter through all the
different modules, only the global lua table is changed, and not the `ConfigurationSet` itself.
To overcome this, a new `LazyConfigurationSet` class is added and is used
instead of `ConfigurationSet` to represent global lua table. The reading of
the table's fields are delayed, and happen only when the set is frozen.