Add support for delayed lua evaluation and configuration edition (#28)
* Store module name and type as a configuration item with special name

* [Docs] document lua namespace

* [Docs] Fix some styling issues

* Add support for lazy lua value

A lazy value is an expression whose evaluation is delayed. In our context,
it essentially means a value which can be changed after the configuration
file is parsed.

An new lua function is added, `parameter` which accepts a string argument.
Instead of referencing directly the variable in the lua configuration, the
user must use this new function, and pass as argument the name of the
parameter, as specified in the `configuration` table (see updated lua
configuration for an example)

* [Docs] move all detailed descriptions to the top of the page

* Move lua initialization to lua/utils.cc

ConfigurationReader is now decoupled from lua/utils via the new
IOnModuleAdded interface.

* Add a way to edit the configuration before freezing

A new method `set()` is added to `ConfigurationSet` which allow the user
to change a parameter *before* freezing the configuration.

In order to correctly dispatch the change of a parameter through all the
different modules, only the global lua table is changed, and not the `ConfigurationSet` itself.
To overcome this, a new `LazyConfigurationSet` class is added and is used
instead of `ConfigurationSet` to represent global lua table. The reading of
the table's fields are delayed, and happen only when the set is frozen.

* [Docs] Update style, and exclude tests