Add looper module to loop over solutions
This is a major change in how the graph is evaluated. Previously, each
module was responsible for looping over all possible solutions, and had
to produce the same number of outputs at the number of inputs, otherwise
the sync is lost. Dropping a solution was almost impossible.

Now, only the Blocks produce a set of solution. All other modules
expected in input only one solution. To deal with multiple solution, a
new module is introduced, the Looper. As its name indicates, it loops
over a set a solution and execute a set of modules for each iteration.

A new concept is introduced: the path. A path is simply a set of modules
that can be passed as input to another module.

A solution class is also added. A solution is simply a set of P4, along
with a jacobian.