Improve handling of solutions with a Looper module (#69)
* Add looper module to loop over solutions

This is a major change in how the graph is evaluated. Previously, each
module was responsible for looping over all possible solutions, and had
to produce the same number of outputs at the number of inputs, otherwise
the sync is lost. Dropping a solution was almost impossible.

Now, only the Blocks produce a set of solution. All other modules
expected in input only one solution. To deal with multiple solution, a
new module is introduced, the Looper. As its name indicates, it loops
over a set a solution and execute a set of modules for each iteration.

A new concept is introduced: the path. A path is simply a set of modules
that can be passed as input to another module.

A solution class is also added. A solution is simply a set of P4, along
with a jacobian.

* Add a way for modules to return status of last evaluation

Modules can now indicate what should be done after its evaluation:
 - Continue evaluation of other modules
 - Stop evaluation and skip to next integration step
 - Stop evaluation and skip to next loop iteration
 - Abort evaluation

* Unit tests for lua Path

* Move all configurations to new Looper syntax

* [Docs] Update documentation of Blocks after renaming of outputs

* Refactor ConfigurationReader / Configuration

* Improve Path handling

Modules now interact only with the Path class, while dirty internal things
are done inside PathElements.

* Decouple Configuration from ParameterSet

* Store the full Configuration::Module inside the pool description

We need at some point to retrieve the parameters of a given module
when building the graph.

* Ensure any module using a looper output is actually in the looper's loop

* Add changelog