Fix configuration freezing when using a Looper
The looper path was not saving a copy of the modules it needs to run,
but used directly the list coming from the configuration file. As soon
as it's modified, after a freeze for example, the list is no longer
valid.

The looper now freeze the path before using it, enforcing a copy of the
list of modules. This is done in `configure`, which is now called
**after** the Graph evaluation, instead of before as previously.